Fold delivery types into utils; retire the handoff env override
delivery-types.ts existed only to break an import cycle between utils
(prompt builders take DeliveryContext) and delivery (needs Decision and
demoteStaleInterrupt). Its four declarations depend on nothing but
utils's own Action type, so they live in utils.ts and the third file
goes away. Docs updated: four root modules.

HYDRA_OBSERVER_HANDOFF forced either observation handoff style on every
provider for A/B reproducibility. The A/B is settled and frozen in the
research branch, no experiment references the variable, and a production
env var that silently changes the model-facing contract is a liability.
usesSplitObservationHandoff now takes only the api.
